requirements.txt
ntscraper==0.3.17
pymongo
python-dotenv